+++ nakadion/CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - metric for `cursor_received_at` in committer
 - `WarnNoFramesSecs` time to emit a warning if no frames were received from Nakadi
 - `WarnNoEventsSecs` time to emit a warning if no events were received from Nakadi
+- added metric to trigger when no events or frames were received and a warning should be emitted
 
 ### REMOVED